server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
python-dotenv==0.17.1
